test(solution): e2e tests for handling circular references
