+++ tests/circular-reference-project/index.spec.ts
@@ -1,0 +1,1 @@
+export { sampleObj } from './index';
