+++ tests/circular-reference-project/index.ts
@@ -1,0 +1,12 @@
+const sampleObj: { [index: string]: any } = {
+    a: {},
+};
+sampleObj.circularReference = sampleObj;
+sampleObj.a.deepCircularReference = sampleObj;
+sampleObj.b = {
+    a: sampleObj.a,
+};
+
+export {
+    sampleObj,
+};
